// src/models/IZenHaDeviceDetails.ts
/* eslint-disable @typescript-eslint/indent */
export interface IZenHaDeviceDetails {
  deviceKey: string;
  deviceName: string;
  enable: boolean;
  ip: string;
  lcnSupport: number;
  online: boolean;
  password: string;
  port: number;
  productKey: string;
  productModel: string;
  protocol: string;
  server: string;
  snNumber: string;
  username: string;
}